% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_1.R
\name{BCa}
\alias{BCa}
\title{BCa confidence interval}
\usage{
BCa(x, alpha, B)
}
\arguments{
\item{x}{a vector composed of random sample points}

\item{alpha}{confidence level}

\item{B}{number of bootstrap replicates}
}
\value{
a named vector representing the confidence interval whose first and second elements are respectively lower bound and upper bound
}
\description{
A function calculating BCa confidence interval for population mean using R
}
\examples{
\dontrun{
x<-rt(200,3)
BCa(x,0.95,500)
}
}
