% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jack}
\alias{jack}
\title{Jackknife estimator for mean and its variance}
\usage{
jack(x)
}
\arguments{
\item{x}{a vector composed of sample points}
}
\value{
a 2-dimension vector represent the jackknife estimate of mean and its variance
}
\description{
Jackknife estimator for mean and its variance using Rcpp
}
\examples{
\dontrun{
x<-rnorm(1000,5,5)
jack(x)
}
}
