% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_1.R
\name{kde}
\alias{kde}
\title{Kernel density estimator}
\usage{
kde(x)
}
\arguments{
\item{x}{a vector composed of random sample points}
}
\value{
a histogram for sample frequency whith the estimated density curve
}
\description{
A kernel density estimator using R
}
\examples{
\dontrun{
x<-rt(2000,3)
kde(x)
}
}
