\name{SC19009-package}
\alias{SC19009-package}
\docType{package}
\title{
  This package is the final for Statistical Computing course consists of two functions and all homework answers.
}
\description{
  Function_1 is a kernel density estimator for single population.
  Function_2 is a jackknife estimator for population mean and its variance.
  Homework answers are all included in the vignette.
}
\details{
  Function_1 is named as kde, input a vector composed of sample points to obtain a histogram and fitted density curve.
  Function_2 is named as jack, input a vector composed of sample points to obtain the jackknife estimate for population mean and its variance (variance estimate for the mean estimate).
}
\author{
SC19009
}
\keyword{ package }