\name{SC19009-package}
\alias{SC19009-package}
\docType{package}
\title{
  This package is the final for 'Statistical Computing' course consists of two functions and all homework answers.
}
\description{
  Function_1 calculates BCa confidence interval for population mean.
  Function_2 is a jackknife estimator for population mean and its variance.
  Homework answers are all included in the vignette.
}
\details{
  Function_1 is named as BCa. Input a vector composed of sample points, confidence level, and bootstrap replicates to obtain a BCa confidence interval for mean.
  Function_2 is named as jack. Input a vector composed of sample points to obtain the jackknife estimate for population mean and its variance (estimated variance for the mean estimator).
}
\author{
SC19009
}
\keyword{ package }